Simplify sphere-config and sphere-serial.
Now we assume that sphere-go-serial and sphere-go-config are in the same
directory as the executing script.

Further, we assume that these scripts are in $(sphere-installation)/bin and
therefore that .. is the location of the sphere installation. So,
if sphere_installDirectory is not set we set it before invoking the tools.